--- wow_simulator_web/simulator_interface/templates/components_new/buffs.html
@@ -1,15 +1,0 @@
-<h3><i>Buffs</i></h3>
-<div class="rounded border" style="padding-left:50px; padding-right:50px; padding-top:20px; padding-bottom:20px; border-color: darkgrey !important;">
-    {% for field in form|filteritemsstartswith:'buffs-' %}
-        {% if forloop.first %}<div class="row">{% endif %}
-
-        <div class="col-sm-4" style="text-align:left">
-            <label class="custom-control custom-checkbox">
-                {{ field }} <span for="{{ field.id }}" class="custom-control-label" > {{ field.label }} </span>
-            </label>
-        </div>
-
-        {% if forloop.counter|divisibleby:3 %}</div><br><div class="row">{% endif %}
-        {% if forloop.last %}</div>{% endif %}
-    {% endfor %}
-</div>
