--- wow_simulator_web/simulator_interface/templates/components_new/weapons.html
@@ -1,18 +1,0 @@
-<h3><i>Weapons</i></h3>
-<div class="rounded border" style="padding-left:50px; padding-right:50px; padding-top:20px; padding-bottom:20px; border-color: darkgrey !important;">
-    {% for field in form|filteritemsstartswith:'weapons-' %}
-        {% if forloop.first %}<div class="row">{% endif %}
-        <div class="col-sm-6">
-            <label for="drop_weapons-{{ field.name }}" data-toggle="tooltip" data-placement="top" title="{{ field.label }} Items"><b>{{ field.label }}</b></label>
-            {{ field }}
-            <br>
-            <br>
-            {% for enchants in form|filteritemsstartswith:'weaponsenchants-'|filteritemsendswith:field.label %}
-                {{ enchants }}
-            {% endfor %}
-        </div>
-        {% if forloop.counter|divisibleby:2 %}</div><br><br><div class="row">{% endif %}
-        {% if forloop.last %}</div>{% endif %}
-    {% endfor %}
-</div>
-
